backend/requirements.txt
asgiref
Django~=5.1.6
djangorestframework
django-cors-headers
sqlparse
python-dotenv
